multi process example and update for new simulator command that only samples some sensors
