+++ examples/config/scenario_ego_on_shoulder_straightaway.json
@@ -1,48 +1,55 @@
 {
-    "trigger_boxes": [],
-    "vehicles": [
-        {
-            "body_color": "Carpaint_White",
-            "class_path": "/Game/Vehicles/subcompact_monoDrive_01.subcompact_monoDrive_01_C",
-            "delayed_start": 0.0,
-            "desired_speed": 0.0,
-            "destroy_on_collision": false,
-            "distance_to_stop_sign": 400.0,
-            "ignore_lane_change_boxes": false,
-            "ignore_other_vehicles": false,
-            "ignore_traffic_signals": false,
-            "initial_speed": 0.0,
-            "lane_offset": 0.0,
-            "limit_speed_in_curves": true,
-            "location": {
-                "x": 7210.0,
-                "y": 450.0,
-                "z": 30.0
-            },
-            "obey_road_speed_limits": true,
-            "overtake_slower_vehicles": false,
-            "randomize_turn_preferences": true,
-            "rolling_stop_speed": 0.00100000004749745,
-            "rotation": {
-                "pitch": 0.0,
-                "roll": 0.0,
-                "yaw": 0.0
-            },
-            "scale": {
-                "x": 1.0,
-                "y": 1.0,
-                "z": 1.0
-            },
-            "search_forward_distance": 10000.0,
-            "snap_to_lane": true,
-            "steer_forward_distance": 550.0,
-            "tags": [
-                "vehicle",
-                "dynamic",
-                "car",
-                "ego"
-            ],
-            "type": "subcompact_monoDrive_01_C"
-        }
-    ]
+  "trigger_boxes": [],
+  "vehicles": [
+    {
+      "body_color": "Carpaint_White",
+      "class_path": "/Game/Vehicles/subcompact_monoDrive_01.subcompact_monoDrive_01_C",
+      "delayed_start": 0.0,
+      "desired_speed": 0.0,
+      "destroy_on_collision": false,
+      "distance_to_stop_sign": 400.0,
+      "ignore_lane_change_boxes": false,
+      "ignore_other_vehicles": false,
+      "ignore_traffic_signals": false,
+      "initial_speed": 0.0,
+      "initial_speed_is_delta": true,
+      "lane_offset": 0.0,
+      "limit_speed_in_curves": true,
+      "location": {
+        "x": 7210.0,
+        "y": 450.0,
+        "z": 30.0
+      },
+      "name": "subcompact_monoDrive_01_C_1",
+      "obey_road_speed_limits": true,
+      "overtake_slower_vehicles": false,
+      "randomize_turn_preferences": true,
+      "rolling_stop_speed": 0.00100000004749745,
+      "rotation": {
+        "pitch": 0.0,
+        "roll": 0.0,
+        "yaw": 0.0
+      },
+      "scale": {
+        "x": 1.0,
+        "y": 1.0,
+        "z": 1.0
+      },
+      "search_forward_distance": 10000.0,
+      "snap_to_lane": true,
+      "spawn_offset": {
+        "x": 0.0,
+        "y": 0.0,
+        "z": 0.0
+      },
+      "steer_forward_distance": 550.0,
+      "steer_forward_distance_lane_change": 1500.0,
+      "tags": [
+        "vehicle",
+        "dynamic",
+        "car",
+        "ego"
+      ]
+    }
+  ]
 }
