+++ examples/cpp/CMakeLists.txt
@@ -12,7 +12,6 @@ if(EIGEN3_INCLUDE_DIR)
         add_subdirectory(lane_follower)
         add_subdirectory(dev)
         add_subdirectory(multi_process_example)
-        add_subdirectory(multi_process_example_new_command)
     else(${OPENCV_FOUND})
         message("Warning, lane_follower and dev examples not built, missing Opencv.")
     endif(${OPENCV_FOUND})
