+++ examples/cpp/multi_process_example/CMakeLists.txt
@@ -6,17 +6,19 @@ set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-add_executable(simulator_controller simulator_controller.cpp)
-add_executable(vehicle_controller vehicle_controller.cpp)
+add_executable(sim_control simulator_controller.cpp)
+add_executable(vehicle_control vehicle_controller.cpp)
+add_executable(video_display video_display.cpp)
 
-target_link_libraries(simulator_controller monodrive)
-target_link_libraries(vehicle_controller monodrive)
+target_link_libraries(sim_control monodrive)
+target_link_libraries(vehicle_control monodrive)
+target_link_libraries(video_display monodrive)
 
 include_directories(${EIGEN3_INCLUDE_DIR})
 
 find_package(OpenCV REQUIRED PATHS "C:/local/opencv/build" "C:/opencv/build")
 include_directories(${OpenCV_INCLUDE_DIRS})
 
-target_include_directories(vehicle_controller PUBLIC ${OpenCV_INCLUDE_DIRS})
-link_libraries(vehicle_controller ${OpenCV_LIBRARIES})
-target_link_libraries(vehicle_controller ${OpenCV_LIBRARIES})
+target_include_directories(video_display PUBLIC ${OpenCV_INCLUDE_DIRS})
+link_libraries(video_display ${OpenCV_LIBRARIES})
+target_link_libraries(video_display ${OpenCV_LIBRARIES})
